pyproject.toml
[tool.poetry]
name = "bambi"
version = "0.1.0"
description = "Background Ambient Lighting for your monitor"
authors = ["Harry Peach"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.1.2"
colored = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylama = "^7.7.1"
autopep8 = "^1.5.5"
expects = "^0.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
